create newrace module and update utils.js file with serializeform function
