+++ public/app.js
@@ -4,6 +4,7 @@ import navbar from './navbar.js'
 import calendrier from './calendrier.js';
 import formulaire from './formulaire.js'
 import admin from './admin.js'
+import newRace from './newrace.js'
 const mainDiv = document.getElementById('main')
  
 const render = html => {
@@ -36,7 +37,7 @@ const controllers = {
       `)
     ),
     '/calendrier': calendrier
-      ,
+    ,
     '/classement': () =>
       fetch('/courses')
       .then(res => res.json())
@@ -82,47 +83,7 @@ const controllers = {
       }
     ),
 
-    '/race/new': () => {
-      render(`
-          ${navbar}
-          <div class="container">
-            <div id="alert-box" class="hidden">
-          </div>
-          <h2>Création de la course</h2>
-          <form id="add-race">
-          <div class="form-group">
-          <label for="inputdate">Date de la course (YYYY-MM-DD HH:MM:SS.SSS)</label>
-            <input name="date" type="text" class="form-control" id="inputFirstName" placeholder="YYYY-MM-DD HH:MM:SS.SSS">
-          </div>
-          <button type="submit" class="btn btn-primary">Créer votre course</button>
-        </form>
-      </div>
-      `
-      )
-      //js du formulaire création de course
-      const form = document.getElementById('add-race')
-      form.addEventListener('submit', e => {
-        e.preventDefault()
-        const data = serializeForm(form)
-        fetch('/race', {
-          method: 'POST',
-          headers: {
-            'Accept': 'application/json, text/plain, */*',
-            'Content-Type': 'application/json'
-          },
-          body: JSON.stringify(data)
-        })//appel cette fonction pour gérer les routes
-        .then(res => res.json())
-        .then(races => {
-          const alertBox = document.getElementById('alert-box')
-          alertBox.className = 'alert alert-success'
-          alertBox.innerHTML += `Course créée`
-         // alertBox.innerHTML += `\n Vous allez être redirigés vers la page d'acceuil`
-        })
-        window.setTimeout(() =>
-        { window.location = "/"; },3000);
-      })
-    },
+    '/race/new': newRace,
 
     '/admin': () => {
       render(`${admin}`)
