+++ public/calendrier.js
@@ -83,5 +83,4 @@ Launch demo modal
   }
 })
 
-
 export default calendrier
