+++ public/utils.js
@@ -14,3 +14,12 @@ const mainDiv = document.getElementById('main')
 export const render = html => {
     mainDiv.innerHTML = html
   }
+
+  export const serializeForm = form => {
+    const data = {}
+    const elements = form.getElementsByClassName('form-control')
+    for(let el of elements) {
+      data[el.name] = el.value
+    }
+    return data
+  }
